Managed problem indexes using shared prefs.
I'm not satisfied by this solution.  At this moment, every 24 hours,
assuming that the problem data will be updated automatically, I reset
the problem index to 0 so that latest problems are asked prioritarily.

However, as I don't keep a history of all the asked problems, it means
that the same problems will be asked more than once. For example, let's
say that on day 1, there are 5 problems asked: a, b, c, d, e.  On day 2,
as new problems are inserted in the database, we will ask the latest
problems first: 1, 2, 3.  After that the 3 latest problems have been
asked, I continue by asking again a, b, c, d, e.  This is not very
nice.

I think the best solution will be for the server to accumulate the problem ids that have been
asked.  The server will then be able to make sure that the same problem
is never asked twice.