Changed kanji selection algo and show additional kanji pages.
Two sets of kanjis are computed. The first set is the kanjis with the
same stroke count as the drawn kanji.  The second set is the kanjis with stroke count
equal to plus or minus one stroke count of the drawn kanji. The two sets
are then merged together. The mixed list of kanjis is shown to the user.