Specified width on all columns.
I did this in order to fix a rendering issue that occurs only when I run
the application on the Nexus Galaxy through the emulator.  When a
problem has a dead link, the row is shorter than the others.
Unfortunately, this doesn't fix the problem.

As this problem doesn't occur on my tablet, I suspect it's an issue with
the emulator.