Regenerated strokes files to remove katakanas.
All katakanas and some kanjis that look like katakanas have been
removed to prevent confusion in such a situation where the user could
have 2 characters that look alike.

For example, in a Writing Problem,
if we have to write the word "mouth" ("kuchi").  The katakana "ro" and
the kanji "kuchi" were displayed.  It's not obvious which character
to click because they look identical.