Consider only higher value of the range.
Instead of considering both kanjis with fewer strokeѕ and kanjis with more
strokes, consider only kanjis with more strokes.  This makes the search
faster because it divides the search space by 2.  Practically, users
usually writes kanjis using fewer strokes because the use shortcuts.
Because of this, it's not needed to consider kanjis with fewer strokes.