Changed font for Writing Problems.
Use the guate-luminescence font
(https://www.freejapanesefont.com/gyate-luminescence/).
This font is free to use for non-commercial projects.
It seems to contain all the kanjis and they are drawn correctly (i.e. in
such a way that we can count the strokes properly).　For example, the
kanjis koromo and kita look perfect.