Started implementing background music.
It's far from over but it gives a good idea.
Some code should be discarded as it's not used.
Also, all the other activities should share the Action Bar and its
behavior.
Most likely that the playerAdapter instance and its related data (like isBackgroundMusicOn and backgroundMusicPos) should be stored in
KankenApplication.
At this moment, on this version, the music stops playing when the
application is not active.  To make the music keep playing even while
the user uses a different application, it's recommended to use a service
instead.  I'm not sure if it's a good idea for now.