Show No data for non-archived problems.
As we started to archive the problems since October, this issue should
become less and less relevant but for now, there are still some problems
that are shown in the stats that have not been archived.  For these
problems, the label No data is shown.

Eventually, all the problems will be archived and this won't be
necessary anymore.