Fixed Cannot quit bug.
The problem was that when I call firstActivity.finishAndRemoveTask(), I
did not set firstActivity to null. The next time that the application is
launched, for some reasons, the KankenApplication instance is not
destroyed (memory leak?) and it remembers the value of firstActivity
that is obsolete.  And because of the test in setFirstActivity(), it's
not updated with the latest value. When the user quits the application,
it tries to finish the obsolete activity and that doesn't do anything so
it was impossible to quit the application.