Fixed Activity Stack issue.
When the commit 1ed3081 was introduced, it fixed a Back Button issue but it also
created a new one: When a link is clicked from the Summary page, if we
click Back to come back to the Summary, we ended up in the Settings
page.  Actually, just clicking the Home button from the Summary page
provoked the problem.  When the app is clicked, the Summary page is
gone.  The Settings are shown instead.

I reverted the commit 1ed3081 and made sure that when we go back to
Settings that the flag Intent.FLAG_ACTIVITY_CLEAR_TOP is used.