Fixed another Cannot quit bug.
This was similar to #cf9c38c but a little bit different.
If the user starts a Reading Quiz, then clicks Back, then
starts a Writing Quiz.  The app cannot quit event if we click the Quit
button.

The method setFirstActivity() was called when the second quiz is started but as the firstActivity was not null, the test prevented the firstActivity to be actualized.  So when the Quit button is clicked, the previous and obsolete firstActivity instance was terminated but it had no effect.

I removed the test so that firstActivity is overwritten each time that
setFirstActivity() is called.