Save announcement as preferences instead.
The idea is that if the announcement is about service unavailability, it
would be nice if we could show the announcement without depending on the
server.  If the announcement is stored on the device, it will be
possible to display it when the application starts without performing a
query to the server.